# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gestao-financeira_backend_internal_transaction_application_dtos.CreateTransactionInput:
    properties:
      account_id:
        type: string
      amount:
        type: number
      currency:
        enum:
        - BRL
        - USD
        - EUR
        type: string
      date:
        description: 'ISO 8601 format: YYYY-MM-DD'
        type: string
      description:
        maxLength: 500
        minLength: 3
        type: string
      type:
        enum:
        - INCOME
        - EXPENSE
        type: string
      user_id:
        type: string
    required:
    - account_id
    - amount
    - currency
    - date
    - description
    - type
    - user_id
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.CreateTransactionOutput:
    properties:
      account_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.DeleteTransactionOutput:
    properties:
      message:
        type: string
      transaction_id:
        type: string
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.GetTransactionOutput:
    properties:
      account_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.ListTransactionsOutput:
    properties:
      count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.TransactionOutput'
        type: array
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.TransactionOutput:
    properties:
      account_id:
        type: string
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.UpdateTransactionInput:
    properties:
      amount:
        type: number
      currency:
        enum:
        - BRL
        - USD
        - EUR
        type: string
      date:
        description: 'ISO 8601 format: YYYY-MM-DD'
        type: string
      description:
        maxLength: 500
        minLength: 3
        type: string
      transaction_id:
        type: string
      type:
        enum:
        - INCOME
        - EXPENSE
        type: string
    required:
    - transaction_id
    type: object
  gestao-financeira_backend_internal_transaction_application_dtos.UpdateTransactionOutput:
    properties:
      account_id:
        type: string
      amount:
        type: number
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@gestaofinanceira.com
    name: API Support
  description: API REST para gestão financeira pessoal e profissional. Gerencia usuários,
    contas e transações financeiras.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gestão Financeira API
  version: "1.0"
paths:
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Lists all transactions for the authenticated user. Optionally filter
        by account_id and/or type (INCOME or EXPENSE).
      parameters:
      - description: Filter by account ID (UUID)
        in: query
        name: account_id
        type: string
      - description: Filter by transaction type (INCOME or EXPENSE)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions with count
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.ListTransactionsOutput'
        "400":
          description: Bad request - invalid user ID, account ID or type
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: List transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Creates a new transaction (INCOME or EXPENSE) for the authenticated
        user. Links transaction to an account.
      parameters:
      - description: Transaction creation data (account_id, type, amount, currency,
          description, date)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.CreateTransactionInput'
      produces:
      - application/json
      responses:
        "201":
          description: Transaction data
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.CreateTransactionOutput'
        "400":
          description: Bad request - invalid input data or account not found
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a transaction by its ID (soft delete). The transaction
        is marked as deleted but not permanently removed from the database.
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.DeleteTransactionOutput'
        "400":
          description: Bad request - invalid transaction ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - transaction does not exist
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a transaction
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: Retrieves a specific transaction by its ID. Only returns transactions
        that belong to the authenticated user.
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Transaction data
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.GetTransactionOutput'
        "400":
          description: Bad request - invalid transaction ID format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - transaction does not belong to user
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - transaction does not exist
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get transaction by ID
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Updates an existing transaction. Supports partial updates - only
        provided fields will be updated. At least one field must be provided.
      parameters:
      - description: Transaction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: 'Transaction update data (all fields optional: type, amount,
          currency, description, date)'
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.UpdateTransactionInput'
      produces:
      - application/json
      responses:
        "200":
          description: Updated transaction data
          schema:
            $ref: '#/definitions/gestao-financeira_backend_internal_transaction_application_dtos.UpdateTransactionOutput'
        "400":
          description: Bad request - invalid transaction ID, invalid data, or no fields
            provided
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - transaction does not belong to user
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found - transaction does not exist
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a transaction
      tags:
      - transactions
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"